Extract populate and get state to separate file
